removed practice api, moved endpoints to endpoints directory
