+++ server/app/api/tests/test_app.py
@@ -11,23 +11,6 @@ def test_default():
     assert response.status_code == 200
     assert response.json() == {"message": "Welcome to the Nevermore API"}
 
-def test_add():
-    response = client.post("/ben/add", json={"name": "Black", "color": "#000000"})
-    assert response.status_code == 200
-    data = response.json()
-
-    assert "id" in data
-    assert isinstance(data["id"], int)
-    assert data["name"]== "Black"
-    assert data["hex"]== "#000000"
-
-def test_filter():
-    response = client.post("/ben/filter", json={"name": "red"})
-    assert response.status_code == 200
-
-    matches = response.json().get("filtered", [])
-    for color in matches:
-         assert "red" in color["name"].lower()
 
 def test_add_joke():
     response = client.post("/will/post", json={"typeJoke": "general", "setup": "test setup", "punchline": "test punchline"})
