+++ server/app/models/request_models.py
@@ -10,10 +10,3 @@ class RequestJoke(BaseModel):
     typeJoke: str
     setup: str
     punchline: str
-
-class RequestColor(BaseModel):
-    name: str
-    color: str
-
-class RequestName(BaseModel):
-    name: str
