+++ server/main.py
@@ -1,22 +1,16 @@
 from app.core.application import app
+from app.models.request_models import RequestJoke
+from app.api import ben_endpoints
+
 import httpx
-from app.models.request_models import RequestJoke, RequestColor, RequestName, PromptRequest, PromptResponse
-from fastapi import HTTPException
-from BenOpenAI import get_openai_response
 
 if __name__ == "__main__":
     import uvicorn
     uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)
 
 
+app.include_router(ben_endpoints.router, prefix="/ben")
 
-@app.post("/generate", response_model=PromptResponse)
-async def generate_response(request: PromptRequest):
-    try:
-        response_text = get_openai_response(request.prompt)
-        return PromptResponse(response=response_text)
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
 
 @app.post("/will/post")
 async def will_post(data: RequestJoke):
@@ -32,24 +26,3 @@ async def will_post(data: RequestJoke):
 def read_root():
     return {"message": "Welcome to the Nevermore API"}
 
-@app.post("/ben/filter")
-async def ben_filter(data: RequestName):
-    url = "https://api.sampleapis.com/csscolornames/colors"
-    async with httpx.AsyncClient(timeout=10) as client:
-        response = await client.get(url)
-        response.raise_for_status()
-
-    colors = response.json()
-    search_term = data.name.lower()
-    filtered = [c for c in colors if search_term in c.get("name", "").lower()]
-    
-    return {'filtered': filtered}
-
-@app.post("/ben/add")
-async def ben_add(data: RequestColor):
-    url = "https://api.sampleapis.com/csscolornames/colors" 
-    async with httpx.AsyncClient() as client:
-        response = await client.post(url, json={"name": data.name, "hex": data.color})
-        response.raise_for_status() 
-
-    return response.json()
